pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-ai-coach"
version = "0.1.0"
description = "AI-powered Ironman triathlon coach using Garmin Connect data"
requires-python = ">=3.11"
dependencies = [
    "garminconnect",
    "typer[all]",
    "rich",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml",
    "keyring",
]

[project.scripts]
coach = "coach.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
